requirements.txt
numpy==1.19.3
opencv-python==4.5.1.48
tensorflow==2.4.0

Pillow==8.1.0
lz4==3.1.3
psutil==5.8.0
matplotlib==3.3.3
GPUtil==1.4.0